+++ CHANGELOG.md
@@ -13,6 +13,7 @@
  - Lua: Add iracing_send_current_session() that will make IRacingPlugin send CurrentSession
  - Lua: Add iracing_send_session_state() that will make IRacingPlugin send SessionState
  - Lua: Add iracing_send_race_flags() that will make IRacingPlugin send race flags
+ - InternalInitialized event removed. Refactored to avoid having it
 
 ## [0.3.0](https://github.com/dennis/slipstream/releases/tag/v0.3.0) (2020-01-05)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.2.0...v0.3.0)